Tests/LibWeb/Text/expected/HTML/button-attributes.txt

submit
button
button
submit
submit
